Implement Dornheim 2019 (Sign problem) and 2023 (Fictitious sign) figs.
Still some kind of offset (?) error in the estimators.
MC VERY badly equilibrirated - strong finite size effet with the box
size. Something to do with the Worm-style PBC moves from Spada not
interacting well with a confining potential. We initialise distributed
across the box at random, and then fail to get confined.
So I'm not going to spend too much time chasing terms in the estimators
(which seem to reproduce Spada perfectly now... though Spada has no
external potential), if it might just be we're massivelu out of eqm for
the sampling.